admin/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Lamella Luciana Elizabeth <lulilamella@gmail.com>"]
readme = "README.md"
packages = [{include = "web", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.3.3"
psycopg2-binary = "^2.9.7"
flask-sqlalchemy = "^3.1.1"
flask-mail = "^0.9.1"
marshmallow = "^3.20.1"
authlib = "^1.2.1"
requests = "^2.31.0"
pyopenssl = "^23.3.0"
flask-cors = "^4.0.0"
flask-jwt-extended = "^4.5.3"
google-auth = "^2.0"
google-auth-oauthlib = "^0.4"
flask-login = "^0.6.3"
flask-paginate = "^2023.10.24"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
